requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
beartype==0.21.0
catalogue==2.0.10
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
confection==0.1.5
contourpy==1.0.7
cycler==0.11.0
decorator==5.2.1
deepdiff==8.6.1
dill==0.4.0
distlib==0.4.0
executing==2.2.0
face-recognition-models==0.3.0
fastapi==0.117.1
fastcore==1.8.8
filelock==3.13.1
fonttools==4.39.3
fsspec==2024.6.1
fuzzywuzzy==0.18.0
h11==0.16.0
icecream==2.1.8
identify==2.6.14
idna==3.10
imageio==2.37.0
imgaug==0.4.0
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
jsonlines==4.0.0
kiwisolver==1.4.4
lazy_loader==0.4
Levenshtein==0.27.1
loguru==0.7.3
lovely-numpy==0.2.13
lovely-tensors==0.1.18
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
orderly-set==5.5.0
packaging==23.0
pandas==2.3.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.4.0
plum-dispatch==2.5.7
polars==1.33.1
pre_commit==4.3.0
prompt_toolkit==3.0.50
psutil==7.1.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
rich==14.1.0
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
srsly==2.5.1
stack-data==0.6.3
starlette==0.48.0
sudo==1.0.0
sympy==1.13.3
tifffile==2025.9.20
torch_snippets==0.560
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.1
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.36.0
virtualenv==20.34.0
wasabi==1.1.3
wcwidth==0.2.13
win32_setctime==1.2.0
xmltodict==1.0.2
